+++ Quadratures.h
@@ -8,6 +8,6 @@
 
 using namespace std;
 
-vector<long double> GaussQuadratureWeights(vector<long double> & );
+vector<double> GaussQuadratureWeights(vector<double> & );
 
-void GaussQuadratureNodes ( int , vector<long double> &, vector<long double> &);
+void GaussQuadratureNodes ( int , vector<double> &, vector<double> &);
